do check before delete a list

+++ src/main/webapp/jsp/usr/article/list.jsp
@@ -24,7 +24,8 @@ List<ArticleDto> articles = (List<ArticleDto>)request.getAttribute("articles");
 
                    <!-- flex-grow : 성장성 -->
                 <a class="flex-grow hover:underline hover:text-[red]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getTitle()%></a>
-                <a class="w-[100px] hover:underline hover:text-[red]" href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
+
+                <a onclick="if ( !confirm('정말로 삭제하시겠습니까?') ) return false;" class="w-[100px] hover:underline hover:text-[red]" href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
 
             </li>
         <% } %>
